- Convert status cli as a subcommand of crio - Moved commands to the internal criocli package and used them in the crio and crio-status binaries - Adding the status subcommands directly to the  binary
