+++ completions/fish/crio-status.fish
@@ -24,11 +24,11 @@ complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'man' -d 'Ge
 complete -c crio-status -n '__fish_seen_subcommand_from markdown md' -f -l help -s h -d 'show help'
 complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'markdown md' -d 'Generate the markdown documentation.'
 complete -c crio-status -n '__fish_seen_subcommand_from config c' -f -l help -s h -d 'show help'
-complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'config c' -d 'Show the configuration of CRI-O as TOML string.'
+complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'config c' -d 'Show the configuration of CRI-O as a TOML string.'
 complete -c crio-status -n '__fish_seen_subcommand_from containers container cs s' -f -l help -s h -d 'show help'
 complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'containers container cs s' -d 'Display detailed information about the provided container ID.'
 complete -c crio-status -n '__fish_seen_subcommand_from containers container cs s' -f -l id -s i -r -d 'the container ID'
 complete -c crio-status -n '__fish_seen_subcommand_from info i' -f -l help -s h -d 'show help'
-complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'info i' -d 'Retrieve generic information about CRI-O, like the cgroup and storage driver.'
+complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'info i' -d 'Retrieve generic information about CRI-O, such as the cgroup and storage driver.'
 complete -c crio-status -n '__fish_seen_subcommand_from help h' -f -l help -s h -d 'show help'
 complete -r -c crio-status -n '__fish_crio-status_no_subcommand' -a 'help h' -d 'Shows a list of commands or help for one command'
