+++ completions/zsh/_crio
@@ -12,6 +12,7 @@ by CRI-O. This allows you to save you current configuration setup and then load
 it later with **--config**. Global options will modify the output.'
         'version:display detailed version information'
         "wipe:wipe CRI-O's container and image storage"
+        'status:Display status information'
         'help:Shows a list of commands or help for one command'
         'h:Shows a list of commands or help for one command'
   )
