+++ completions/zsh/_crio-status
@@ -7,14 +7,14 @@ _cli_zsh_autocomplete() {
         'man:Generate the man page documentation.'
         'markdown:Generate the markdown documentation.'
         'md:Generate the markdown documentation.'
-        'config:Show the configuration of CRI-O as TOML string.'
-        'c:Show the configuration of CRI-O as TOML string.'
+        'config:Show the configuration of CRI-O as a TOML string.'
+        'c:Show the configuration of CRI-O as a TOML string.'
         'containers:Display detailed information about the provided container ID.'
         'container:Display detailed information about the provided container ID.'
         'cs:Display detailed information about the provided container ID.'
         's:Display detailed information about the provided container ID.'
-        'info:Retrieve generic information about CRI-O, like the cgroup and storage driver.'
-        'i:Retrieve generic information about CRI-O, like the cgroup and storage driver.'
+        'info:Retrieve generic information about CRI-O, such as the cgroup and storage driver.'
+        'i:Retrieve generic information about CRI-O, such as the cgroup and storage driver.'
         'help:Shows a list of commands or help for one command'
         'h:Shows a list of commands or help for one command'
   )
